refactor(console-ui): phase 5a — api-console feature module (F7/F6/F8)
Split the 543-line api-console page (~310 lines of static data) into a module:
- content.ts: the ENDPOINTS reference + Endpoint/CodeSnippet types + the SDK
  snippet builders (sdkSetupExamples/chatExamples/modelsExamples). Pure data.
- EndpointRow.tsx: the one interactive island (expand + analytics).
- page.tsx: thin orchestrator (543 -> ~145 lines) wired to STORAGE_KEYS +
  clientCoordinatorUrl (drops its private storage-key/coordinator-URL constants,
  F6). Extracting the static content sets up an RSC conversion later (F8).

Invariant held: build green, eslint 0 errors / 112 warnings, vitest 358 pass /
4 pre-existing fail.
